made constrained seperate from successor generator

+++ include/PetriEngine/ExplicitColored/AtomicTypes.h
@@ -12,7 +12,7 @@ namespace PetriEngine::ExplicitColored {
     using Variable_t = uint32_t;
     using MarkingCount_t = uint32_t;
     using sMarkingCount_t = int32_t;
-    using Binding_t = uint32_t;
+    using Binding_t = uint64_t;
     constexpr Color_t DOT_COLOR = 0;
     constexpr Color_t ALL_COLOR = std::numeric_limits<Color_t>::max();
 }
